+++ src/methods/$join/async-join.mjs
@@ -6,10 +6,10 @@
  * More information can be found in CONTRIBUTING.md
  */
 
-import { asyncIterableCurry } from '../../internal/async-iterable';
+import { asyncWrapWithMethodIterable, asyncEnsureIterable } from '../../internal/async-iterable';
 import { asyncJoinWithSubseq } from '../$join-with-subseq/async-join-with-subseq';
 const emptySubseq = [];
 export function asyncJoin(source) {
-  return asyncJoinWithSubseq(source, emptySubseq);
+  return asyncJoinWithSubseq(asyncEnsureIterable(source), emptySubseq);
 }
-export default asyncIterableCurry(asyncJoin);
+export default asyncWrapWithMethodIterable(asyncJoin);